Add files via upload
Sorry but I cannot export to DXF or DWG from Fusion 360. I guess that is because they are direct competitors. I saved as a STEP because most free CAD/CAM systems can read them and they are much easier to work with that an STL file. Feel free to use these as you see fit as I am offering this as open source for personal use. My files are not available for commercial purposes without a licensing agreement.